Ajout de restrictions dans ForumController

+++ controller/ForumController.php
@@ -74,7 +74,8 @@ class ForumController extends AbstractController implements ControllerInterface
     public function addCategorie()
     {
         // Si on clique sur le bouton "Ajouter une catégorie"
-        if (isset($_POST["addCategorie"])) {
+        if (isset($_POST["addCategorie"]) && (Session::isAdmin())) {
+
             // On récupère le nom de la catégorie et on le filtre
             $nomCategorie = filter_input(INPUT_POST, 'nomCategorie', FILTER_SANITIZE_SPECIAL_CHARS);
             // On instancie le manager des catégories
@@ -87,40 +88,49 @@ class ForumController extends AbstractController implements ControllerInterface
             // On ajoute la catégorie
             $categorieManager->add($data);
             // On redirige vers la liste des catégories
-            $this->redirectTo("forum", "listCategories");
+        } else {
+            Session::addFlash("error", "Vérifiez que vous êtes bien connectés !");
         }
+        $this->redirectTo("forum", "listCategories");
     }
 
     // Supprimer une catégorie
     public function deleteCategorie($id)
     {
-        // On instancie le manager des catégories
-        $categorieManager = new CategorieManager();
-        // On supprime la catégorie par son id
-        $categorieManager->delete($id);
-        // On redirige vers la liste des catégories
+        if ((Session::isAdmin())) {
+            // On instancie le manager des catégories
+            $categorieManager = new CategorieManager();
+            // On supprime la catégorie par son id
+            $categorieManager->delete($id);
+            // On redirige vers la liste des catégories
+        } else {
+            Session::addFlash("error", "Vous n'avez pas les droits pour cet action!");
+        }
         $this->redirectTo("forum", "listCategories");
     }
+
     // Modifier une catégorie
     public function updateCategorie($id)
     {
-        if (isset($_POST["updateCategorie"]) && isset($_POST["nomCategorie"]) && !empty($_POST["nomCategorie"])) {
-            $nomCategorie = filter_input(INPUT_POST, 'nomCategorie', FILTER_SANITIZE_SPECIAL_CHARS);
+        if (isset($_POST["updateCategorie"]) && isset($_POST["nomCategorie"]) && !empty($_POST["nomCategorie"]) && (Session::isAdmin())) {
+            $nomCategorie = filter_input(INPUT_POST, 'nomCategorie', FILTER_SANITIZE_FULL_SPECIAL_CHARS);
             if ($nomCategorie) {
                 $categorieManager = new CategorieManager();
                 $categorieManager->updateCategorie($id, $nomCategorie);
                 $this->redirectTo("forum", "listCategories");
             }
+        } else {
+            Session::addFlash("error", "Vous n'avez pas les droits pour cet action!");
         }
     }
 
     // Ajouter un post
     public function addPost($id)
     {
-        // Si on clique sur le bouton "Ajouter un post" alors on récupère le texte du post et on le filtre
-        if (isset($_POST["addPost"])) {
+        // Si on clique sur le bouton "Ajouter un post" alors on récupère le texte du post et on le filtre, ensuite on vérifie si l'utilisateur est connecté ou qu'il est admin
+        if (isset($_POST["addPost"]) && (Session::isAdmin()) || (isset($_SESSION["user"]) && $_SESSION["user"]->getId($id))) {
             // On récupère le texte du post et on le filtre
-            $texte = filter_input(INPUT_POST, 'texte', FILTER_SANITIZE_SPECIAL_CHARS);
+            $texte = filter_input(INPUT_POST, 'texte', FILTER_SANITIZE_FULL_SPECIAL_CHARS);
             $postManager = new PostManager();
 
             // On ajoute le post selon les données suivantes (texte, user_id, topic_id). Le user_id est fixé à 2 car on n'a pas encore de système de connexion
@@ -129,31 +139,35 @@ class ForumController extends AbstractController implements ControllerInterface
                 "user_id" => $_SESSION["user"]->getId(),
                 "topic_id" => $id
             ];
-            // var_dump($data);
-            // die;
             $postManager->add($data);
-            $this->redirectTo("forum", "listPostsByTopic", $id);
+        } else {
+            Session::addFlash("error", "Vous devez être connectés pour faire cet action!");
         }
+        $this->redirectTo("forum", "listPostsByTopic", $id);
     }
 
     // Méthode pour suppimer un post sauf si c'est le premier post du topic avec findFirstPostByTopic()
     public function deletePost($id)
     {
+        // On vérifie si l'utilisateur est admin ou l'auteur du post pour supprimer le post
+        if ((Session::isAdmin()) || ($_SESSION["user"]) && $_SESSION["user"]->getId($id)) {
+            // On instancie les managers
+            $postManager = new PostManager();
 
-        // On instancie les managers
-        $postManager = new PostManager();
+            // On récupère l'id du topic du post
+            $post = $postManager->findOneById($id);
+            $idTopic = $post->getTopic()->getId();
 
-        // On récupère l'id du topic du post
-        $post = $postManager->findOneById($id);
-        $idTopic = $post->getTopic()->getId();
+            // On récupère l'id du premier post du topic
+            $firstPost = $postManager->findFirstPostByTopic($idTopic);
 
-        // On récupère l'id du premier post du topic
-        $firstPost = $postManager->findFirstPostByTopic($idTopic);
-
-        $idFirstPost = $firstPost->getId();
-        // Si l'id du post est différent de l'id du premier post du topic, on supprime le post
-        if ($id != $idFirstPost) {
-            $postManager->delete($id);
+            $idFirstPost = $firstPost->getId();
+            // Si l'id du post est différent de l'id du premier post du topic, on supprime le post
+            if ($id != $idFirstPost) {
+                $postManager->delete($id);
+            }
+        } else {
+            Session::addFlash("error", "Vous devez être connectés pour faire cet action!");
         }
         // On redirige vers la page du topic
         $this->redirectTo("forum", "listPostsByTopic", $idTopic);
@@ -162,29 +176,38 @@ class ForumController extends AbstractController implements ControllerInterface
     // Modifier un post
     public function updatePostForm($id)
     {
-        $postManager = new PostManager();
-        $topicId = $postManager->findOneById($id)->getTopic()->getId();
-        if (isset($_POST["updatePost"]) && isset($_POST["texte"]) && !empty($_POST["texte"])) {
+        // On instancie le manager des posts
+        if (isset($_POST["updatePost"]) && isset($_POST["texte"]) && !empty($_POST["texte"]) && (Session::isAdmin()) || (isset($_SESSION["user"]) && $_SESSION["user"]->getId($id))) {
+            // On récupère le texte du post et on le filtre
             $texte = filter_input(INPUT_POST, 'texte', FILTER_SANITIZE_SPECIAL_CHARS);
+            // On instancie le manager des posts
+            $postManager = new PostManager();
+            // On récupère l'id du topic du post
+            $topicId = $postManager->findOneById($id)->getTopic()->getId();
             if ($texte) {
+                // On instancie le manager des posts
                 $postManager->updatePostAction($id, $texte);
-                $this->redirectTo("forum", "listPostsByTopic", $topicId);
             }
         }
+        $this->redirectTo("forum", "listPostsByTopic", $topicId);
     }
 
     // Ajouter un topic
     public function addTopic($id)
     {
+        $topicManager = new TopicManager();
+        $postManager = new PostManager();
+
         if (
+            // Vérification des champs du form 
             isset($_POST["addTopic"]) && isset($_POST['titre']) && !empty($_POST['titre'])
-            && isset($_POST['texte']) && !empty($_POST['texte'])
+            && isset($_POST['texte']) && !empty($_POST['texte']
+                // partie Session
+                && (isset($_SESSION["user"]) && $_SESSION["user"]->getId($id)))
         ) {
             // On récupère les différents champs du formulaire pour les filtrer
             $titre = filter_input(INPUT_POST, 'titre', FILTER_SANITIZE_SPECIAL_CHARS);
             $texte = filter_input(INPUT_POST, 'texte', FILTER_SANITIZE_SPECIAL_CHARS);
-            $topicManager = new TopicManager();
-            $postManager = new PostManager();
             $data = [
                 "titre" => $titre,
                 "user_id" => $_SESSION["user"]->getId(),
@@ -199,54 +222,68 @@ class ForumController extends AbstractController implements ControllerInterface
                 "topic_id" => $topicId
             ];
             $postManager->add($data2);
-            $this->redirectTo("forum", "listPostsByTopic", $topicId);
         }
+        $this->redirectTo("forum", "listPostsByTopic", $topicId);
     }
 
     // Supprimer un topic
     public function deleteTopic($id)
     {
-        // On instancie le manager des topics
-        $topicManager = new TopicManager();
-        // On récupère le topic par son id
-        $topic = $topicManager->findOneById($id);
-        // // var_dump($topic);
-        // // die;
-        $idCategorie = $topic->getCategorie()->getId();
-        // // On supprime le topic par son id
-        $topicManager->delete($id);
-        // var_dump($id);
-        // die;
+        if (Session::isAdmin() || (isset($_SESSION["user"]) && $_SESSION["user"]->getId($id))) {
+            // On instancie le manager des topics
+            $topicManager = new TopicManager();
+            // On récupère le topic par son id
+            $topic = $topicManager->findOneById($id);
+
+            $idCategorie = $topic->getCategorie()->getId();
+            // // On supprime le topic par son id
+            $topicManager->delete($id);
+            // On redirige vers la liste des topics
+        } else {
+            Session::addFlash("error", "Vous n'avez pas les droits pour cet action!");
+        }
         $this->redirectTo("forum", "listTopicsByCategorie", $idCategorie);
     }
 
     // Modifier un topic
     public function updateTopic($id)
     {
-        $topicManager = new TopicManager();
-        $categorieId = $topicManager->findOneById($id)->getCategorie()->getId();
-        if (isset($_POST["updateTopic"]) && isset($_POST["titre"]) && !empty($_POST["titre"])) {
-            $titre = filter_input(INPUT_POST, 'titre', FILTER_SANITIZE_SPECIAL_CHARS);
-            if ($titre) {
-                $topicManager->updateTopic($id, $titre);
-                $this->redirectTo("forum", "listTopicsByCategorie", $categorieId);
+        $topicManager = new TopicManager(); // On instancie le mananger
+        if (Session::isAdmin() || (isset($_SESSION["user"]) && $_SESSION["user"]->getId() == $topicManager->findOneById($id)->getUser()->getId())) {
+            if (isset($_POST["updateTopic"]) && isset($_POST["titre"]) && !empty($_POST["titre"])) {
+                $categorieId = $topicManager->findOneById($id)->getCategorie()->getId(); // On récupère l'id de la catégorie du topic
+                $titre = filter_input(INPUT_POST, 'titre', FILTER_SANITIZE_SPECIAL_CHARS);
+                if ($titre) {
+                    $topicManager->updateTopic($id, $titre);
+                }
             }
+        } else {
+            Session::addFlash("error", "Vous n'avez pas les droits pour cet action!");
         }
+        $this->redirectTo("forum", "listTopicsByCategorie", $categorieId);
     }
 
     // Méthode pour vérouiller un topic
     public function lockTopic($id)
     {
-        $topicManager = new TopicManager();
-        $topicManager->lockTopic($id);
+        if (Session::isAdmin()) {
+            $topicManager = new TopicManager();
+            $topicManager->lockTopic($id);
+        } else {
+            Session::addFlash("error", "Vous n'avez pas les droits pour cet action!");
+        }
         $this->redirectTo("forum", "listPostsByTopic", $id);
     }
 
     // Méthode pour dévérouiller un topic
     public function unlockTopic($id)
     {
-        $topicManager = new TopicManager();
-        $topicManager->unlockTopic($id);
+        if (Session::isAdmin()) {
+            $topicManager = new TopicManager();
+            $topicManager->unlockTopic($id);
+        } else {
+            Session::addFlash("error", "Vous n'avez pas les droits pour cet action!");
+        }
         $this->redirectTo("forum", "listPostsByTopic", $id);
     }
 }
